Issue SAN Certificate from CSR (verified)
Usage is for a CSR to be externally prepared and copied into the linux CA. There this script will parse out DNS and IP entries and create a config file. This file is tehn used to create a SAN certificate. Name of certificate is taken from first DNS entry in CSR SAN section.